refclock: fix compiler warning on FreeBSD

+++ refclock_pps.c
@@ -147,7 +147,7 @@ static int pps_poll(RCL_Instance instance)
 
   if (seq == pps->last_seq || UTI_IsZeroTimespec(&ts)) {
     DEBUG_LOG("PPS sample ignored seq=%lu ts=%s",
-              seq, UTI_TimespecToString(&ts));
+              (unsigned long)seq, UTI_TimespecToString(&ts));
     return 0;
   }
